requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
python-multipart
pydantic
PyPDF2
python-docx
pandas
chromadb
openai
tiktoken
python-dotenv
sqlalchemy
psycopg2-binary
python-cors
reportlab
openpyxl
